requirements.txt
# Python 3.12 required
fastapi>=0.115.0
shodan>=1.31.0
uvicorn>=0.34.0
pydantic>=2.11.0
httpx>=0.28.0
geopy>=2.4.0
python-dotenv>=1.1.0
jinja2>=3.1.6